+++ src/main/java/com/flightstats/hub/app/HubBindings.java
@@ -7,12 +7,9 @@ import com.fasterxml.jackson.databind.module.SimpleModule;
 import com.flightstats.hub.alert.AlertRunner;
 import com.flightstats.hub.channel.ChannelValidator;
 import com.flightstats.hub.cluster.*;
-import com.flightstats.hub.dao.ChannelService;
-import com.flightstats.hub.dao.ChannelServiceImpl;
 import com.flightstats.hub.group.GroupProcessor;
 import com.flightstats.hub.group.GroupProcessorImpl;
 import com.flightstats.hub.group.GroupValidator;
-import com.flightstats.hub.health.HubHealthCheck;
 import com.flightstats.hub.metrics.HostedGraphiteSender;
 import com.flightstats.hub.metrics.MetricsRunner;
 import com.flightstats.hub.metrics.MetricsSender;
@@ -52,8 +49,7 @@ public class HubBindings extends AbstractModule {
     @Override
     protected void configure() {
         Names.bindProperties(binder(), HubProperties.getProperties());
-        bind(ChannelService.class).to(ChannelServiceImpl.class).asEagerSingleton();
-        bind(HubHealthCheck.class).asEagerSingleton();
+
         bind(HubClusterRegister.class).asEagerSingleton();
         bind(ZooKeeperState.class).asEagerSingleton();
         bind(ReplicatorManager.class).asEagerSingleton();
